requirements.txt
django==3.2.10
django-tables2==2.4.0
django-autocomplete-light
django-tinymce==3.3.0
dal-admin-filters==1.1.0
django-mptt==0.13.4
django-import-export==2.6.1
gunicorn==20.1.0
whitenoise==5.3.0
python-decouple==3.5
psycopg2==2.9.2
python-dateutil==2.8.2
